requirements.txt
pyserial
flake8
python-dotenv
